Move completion item formatting methods to Remote
